WordPuzzle

Description

Supporting functions for solving letter word puzzles.

Notes

Is functions return a boolean:

isXyz :: a -> Bool

Check functions return an Either:

checkXyz :: a -> Either Error a

TODO Update validation to return a List of errors:

validateXyz :: a -> Validation [Error] a

Documentation

data WordPuzzle

Represent parameters required for the puzzle.

makeWordPuzzle :: Int -> String -> FilePath -> Either ValidationError WordPuzzle

Smart constructor for WordPuzzle.

TODO re-write using accumulative validation

One way to accumulate all validation errors into a list rather than failing on the first error found is to use a monad such as the Writer monad (mtl). Here's how the code could look using the Writer monad:

import Control.Monad.Writer

makeWordPuzzle :: Int -> String -> FilePath -> Writer [ValidationError] WordPuzzle
makeWordPuzzle s ls d = do
   when (not (isSize s)) $ tell [InvalidSize s]
   when (not (isLetters ls)) $ tell [InvalidLetters ls]
   return (WordPuzzle s (head ls) ls d)

Now, when calling the function, you can extract the result and the accumulated errors using runWriter:

result :: (WordPuzzle, [ValidationError])
result = runWriter (makeWordPuzzle size letters filePath)

Note that if there are no validation errors, the list of errors will be empty. You can then check the list of errors to determine if the result is valid or not.

See https://github.com/system-f/validation/blob/master/examples/src/Email.hs

checkSize

Arguments

:: Int
size of word to check

-> Either String Int
Left unexpected size or Right size

Check that mandatory value is in the range from 1 to 9.

>>> checkSize 10
Left (InvalidSize 10)

>>> checkSize 1
Right 1

checkLetters

Arguments

:: String
characters to check

-> Either String String
valid lowercase letters

Check that letters are lowercase alphabetic characters.

hasLetters

Arguments

:: String
valid letters

-> String
dictionary word to check

-> Bool
true if dictionary word matches letters

Check if a word contains only characters from a letters list.

If all valid characters are removed from the word, and there are still characters left over, then the word is not valid.

If all valid characters are removed from the word, and the word is empty, then the word is valid.

hasLetters'

Arguments

:: String
valid letters

-> String
dictionary word to check

-> Bool
true if dictionary word matches letters

Check if a word contains only characters from a letters list.

Original version using set difference.

If all valid characters are removed from the word, and there are still characters left over, then the word is not valid.

If all valid characters are removed from the word, and the word is empty, then the word is valid.

solve :: WordPuzzle -> IO ()

Solve word puzzle given a dictionary of words.

Where each word:

must be greater than the minimum word length

must be no more than 9 characters long

must contain mandatory character

must contain only valid characters

must not exceed valid character frequency

data ValidationError

Error given on invalid parameter.

Constructors

InvalidSize Int
bad size integer

InvalidLetters String
bad letters

UnexpectedValue String
couldn't parse value