WordPuzzle

Description

Supporting functions for solving letter word puzzles.

Notes

Is functions return a boolean:

isXyz :: a -> Bool

Check functions return an Either:

checkXyz :: a -> Either Error a

TODO Update validation to return a List of errors:

validateXyz :: a -> Validation [Error] a

Documentation

data WordPuzzle

Represent parameters required for the puzzle.

makeWordPuzzle :: Int -> String -> FilePath -> Either ValidationError WordPuzzle

Smart constructor for WordPuzzle. TODO re-write using accumulative validation See https://github.com/system-f/validation/blob/master/examples/src/Email.hs

checkSize

Arguments

:: Int
value to check

-> Either String Int
Left unexpected size or Right size

Check that mandatory value is in the range from 1 to 9.

checkLetters

Arguments

:: String
characters to check

-> Either String String
valid lowercase letters

Check that letters are lowercase alphabetic characters.

hasLetters

Arguments

:: String
valid letters

-> String
dictionary word to check

-> Bool
true if dictionary word matches letters

Check if a word contains only characters from a letters list.

If all valid characters are removed from the word, and there are still characters left over, then the word is not valid.

If all valid characters are removed from the word, and the word is empty, then the word is valid.

solve :: WordPuzzle -> IO ()

Solve word puzzle given a dictionary of words. 1. must be greater than the minimum word length 2. must be no more than 9 characters long 3. must contain mandatory character 4. must contain only valid characters 5. must not exceed valid character frequency

data ValidationError

Error given on invalid parameter.

Constructors

InvalidSize Int
bad size integer

InvalidLetters String
bad letters

UnexpectedValue String
couldn't parse value