WordPuzzle

Description

Supporting functions for solving letter word puzzles.

Documentation

remove :: Char -> ByteString -> ByteString

Remove first occurrence of a character from a word.

isValid :: String -> ByteString -> Bool

Check if a word contains only characters from a list.

If all valid characters are removed from the word, and there are still characters left over, then the word is not valid.

If all valid characters are removed from the word, and the word is empty, then the word is valid.