WordPuzzle

Description

Supporting functions for solving letter word puzzles.

Notes

Is functions return a boolean:

isXyz :: a -> Bool

Check functions return an Either:

checkXyz :: a -> Either Error a

TODO Validation returns a List of errors:

validateXyz :: a -> Validation [Error] a

Documentation

data WordPuzzle

Represent parameters required for the puzzle.

makeWordPuzzle :: Int -> Char -> String -> FilePath -> Either ValidationError WordPuzzle

Smart constructor for WordPuzzle. TODO re-write using accumulative validation

checkSize

Arguments

:: Int
value to check

-> Either String Int
Left unexpected size or Right size

Check that mandatory value is in the range from 1 to 9.

checkLetters

Arguments

:: String
characters to check

-> Either String String
valid lowercase letters

Check that letters are lowercase alphabetic characters.

checkMandatory

Arguments

:: String
mandatory character to check

-> Either String Char
valid mandatory letter

Check that mandatory letter. TODO Can we use isMandatory here? It will require an extra parameter, which is not available by the parser.

isWord

Arguments

:: String
valid letters

-> String
dictionary word to check

-> Bool
true if dictionary word matches letters

Check if a word contains only characters from a list.

If all valid characters are removed from the word, and there are still characters left over, then the word is not valid.

If all valid characters are removed from the word, and the word is empty, then the word is valid.

solve :: WordPuzzle -> IO ()

Solve word puzzle.

Print words to stdout.

data ValidationError

Error given on invalid parameter.

Constructors

InvalidSize Int
bad size integer

InvalidMandatory Char
bad mandatory character

InvalidLetters String
bad letters

UnexpectedValue String
couldn't parse value