WordPuzzle

Description

Supporting functions for solving letter word puzzles.

Supporting Functions

remove :: Char -> String -> String

Remove first occurrence of a character from the word.

c.f. http://hackage.haskell.org/package/base/docs/Data-List.html#v:delete

isValid :: String -> String -> Bool

Check if a word contains only characters from a list. If all valid characters are removed from the word, and there are still characters left over, then the word is not valid. If all valid characters are removed from the word, and the word is empty, then the word is valid.

See also Data.List intersection function, \\

isPlural :: String -> Bool

Check if word is a plural.

This assumes the word is at least 2 characters long.

False if word does not end in s

False if word ends in "ss"

True Otherwise as ends in single s