WordPuzzle

Description

Supporting functions for solving letter word puzzles.

Documentation

remove

Arguments

:: Char
character to remove

-> ByteString
string to remove character from

-> ByteString
result string with one instance of character removed

Remove first occurrence of a character from a word.

isValid

Arguments

:: String
valid letters

-> ByteString
dictionary word to check

-> Bool
true if dictionary word matches letters

Check if a word contains only characters from a list.

If all valid characters are removed from the word, and there are still characters left over, then the word is not valid.

If all valid characters are removed from the word, and the word is empty, then the word is valid.